automation/scripts/career_automation.py
#!/usr/bin/env python3
"""
Professional Portfolio & Career Development Automation System
=============================================================

A comprehensive automation system for IT/Cybersecurity professionals to:
- Generate automated resumes/CVs from GitHub activity and certifications
- Maintain dynamic portfolio websites with live metrics
- Track certifications and academic progress
- Automate professional networking and job applications
- Provide skills assessment and learning recommendations
- Generate analytics and career insights

Author: Ariff Mohamed
Target: IT/Cybersecurity professionals pursuing career advancement
"""

import os
import sys
import yaml
import logging
import argparse
from datetime import datetime
from pathlib import Path

# Add project root to Python path
project_root = Path(__file__).parent.parent
sys.path.append(str(project_root))

class CareerAutomationSystem:
    """Main orchestrator for the career development automation system."""
    
    def __init__(self, config_path=None):
        """Initialize the automation system with configuration."""
        self.config_path = config_path or "automation/config/config.yaml"
        self.config = self.load_config()
        self.setup_logging()
        
        # Initialize components
        self.github_collector = None
        self.certification_tracker = None
        self.resume_generator = None
        self.portfolio_manager = None
        self.analytics_engine = None
        self.networking_automation = None
        
    def load_config(self):
        """Load configuration from YAML file."""
        try:
            with open(self.config_path, 'r') as file:
                return yaml.safe_load(file)
        except FileNotFoundError:
            logging.error(f"Configuration file not found: {self.config_path}")
            sys.exit(1)
        except yaml.YAMLError as e:
            logging.error(f"Error parsing configuration file: {e}")
            sys.exit(1)
    
    def setup_logging(self):
        """Configure logging for the system."""
        log_level = logging.INFO
        log_format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        
        logging.basicConfig(
            level=log_level,
            format=log_format,
            handlers=[
                logging.FileHandler('automation/logs/career_automation.log'),
                logging.StreamHandler(sys.stdout)
            ]
        )
        
        self.logger = logging.getLogger(__name__)
        self.logger.info("Career Automation System initialized")
    
    def run_data_collection(self):
        """Collect data from various sources (GitHub, Credly, etc.)."""
        self.logger.info("Starting data collection phase...")
        
        # GitHub data collection
        self.collect_github_data()
        
        # Certification data collection
        self.collect_certification_data()
        
        # Academic progress collection
        self.collect_academic_data()
        
        self.logger.info("Data collection completed")
    
    def collect_github_data(self):
        """Collect GitHub activity and project data."""
        self.logger.info("Collecting GitHub data...")
        # Implementation will be added in github_collector.py
        pass
    
    def collect_certification_data(self):
        """Collect certification data from Credly and Microsoft Learn."""
        self.logger.info("Collecting certification data...")
        # Implementation will be added in certification_tracker.py
        pass
    
    def collect_academic_data(self):
        """Collect academic progress and research milestone data."""
        self.logger.info("Collecting academic data...")
        # Implementation will be added in academic_tracker.py
        pass
    
    def generate_resume(self, template_type="professional"):
        """Generate automated resume/CV."""
        self.logger.info(f"Generating resume using {template_type} template...")
        # Implementation will be added in resume_generator.py
        pass
    
    def update_portfolio(self):
        """Update the dynamic portfolio website."""
        self.logger.info("Updating portfolio website...")
        # Implementation will be added in portfolio_manager.py
        pass
    
    def run_analytics(self):
        """Generate analytics and insights."""
        self.logger.info("Running analytics engine...")
        # Implementation will be added in analytics_engine.py
        pass
    
    def automate_networking(self):
        """Handle automated networking tasks (LinkedIn posts, etc.)."""
        self.logger.info("Running networking automation...")
        # Implementation will be added in networking_automation.py
        pass
    
    def track_job_applications(self):
        """Track and manage job applications."""
        self.logger.info("Tracking job applications...")
        # Implementation will be added in job_tracker.py
        pass
    
    def assess_skills(self):
        """Perform skills assessment and generate learning recommendations."""
        self.logger.info("Running skills assessment...")
        # Implementation will be added in skills_assessor.py
        pass
    
    def run_full_automation(self):
        """Run the complete automation pipeline."""
        self.logger.info("Starting full automation pipeline...")
        
        try:
            # Phase 1: Data Collection
            self.run_data_collection()
            
            # Phase 2: Content Generation
            self.generate_resume()
            self.update_portfolio()
            
            # Phase 3: Analytics and Insights
            self.run_analytics()
            self.assess_skills()
            
            # Phase 4: Automation Tasks
            self.automate_networking()
            self.track_job_applications()
            
            self.logger.info("Full automation pipeline completed successfully")
            
        except Exception as e:
            self.logger.error(f"Error in automation pipeline: {e}")
            raise
    
    def run_component(self, component_name):
        """Run a specific component of the automation system."""
        components = {
            'data_collection': self.run_data_collection,
            'resume': self.generate_resume,
            'portfolio': self.update_portfolio,
            'analytics': self.run_analytics,
            'networking': self.automate_networking,
            'job_tracking': self.track_job_applications,
            'skills': self.assess_skills
        }
        
        if component_name in components:
            self.logger.info(f"Running component: {component_name}")
            components[component_name]()
        else:
            self.logger.error(f"Unknown component: {component_name}")
            self.logger.info(f"Available components: {list(components.keys())}")

def main():
    """Main entry point for the career automation system."""
    parser = argparse.ArgumentParser(
        description="Professional Portfolio & Career Development Automation System"
    )
    
    parser.add_argument(
        '--component',
        choices=['data_collection', 'resume', 'portfolio', 'analytics', 
                'networking', 'job_tracking', 'skills', 'full'],
        default='full',
        help='Component to run (default: full automation)'
    )
    
    parser.add_argument(
        '--config',
        default='automation/config/config.yaml',
        help='Path to configuration file'
    )
    
    parser.add_argument(
        '--template',
        default='professional',
        choices=['academic', 'professional', 'cybersecurity'],
        help='Resume template to use'
    )
    
    args = parser.parse_args()
    
    # Create logs directory if it doesn't exist
    os.makedirs('automation/logs', exist_ok=True)
    
    # Initialize the automation system
    automation_system = CareerAutomationSystem(config_path=args.config)
    
    # Run the requested component
    if args.component == 'full':
        automation_system.run_full_automation()
    elif args.component == 'resume':
        automation_system.generate_resume(template_type=args.template)
    else:
        automation_system.run_component(args.component)

if __name__ == "__main__":
    main()