automation/requirements.txt
# Professional Portfolio & Career Development Automation System
# Requirements for IT/Cybersecurity professionals

# Core dependencies
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping and API interactions
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.9.0

# Resume/CV generation
jinja2>=3.1.0
weasyprint>=60.0
python-docx>=0.8.11

# Academic and research tools
scholarly>=1.7.0
arxiv>=1.4.0

# Social media and networking automation
python-linkedin-v2>=2.0.0
tweepy>=4.14.0

# Database and storage
sqlite3  # Built-in
sqlalchemy>=2.0.0

# Task scheduling and automation
schedule>=1.2.0
celery>=5.3.0
redis>=5.0.0

# Security and encryption
cryptography>=41.0.0
python-jose>=3.3.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0

# Monitoring and logging
structlog>=23.1.0
sentry-sdk>=1.32.0

# LaTeX for resume generation
# Note: Requires separate LaTeX installation (e.g., TeXLive, MiKTeX)

# Optional: Jupyter for interactive development
jupyter>=1.0.0
ipykernel>=6.25.0